requirements.txt
fastapi
uvicorn
requests
python-dotenv
groq
authlib
pyjwt